Changed one to one delete behavior, placeholder bug for arrays (ex:json), added payload in refresh action

+++ src/Database/ModelManager.php
@@ -93,7 +93,9 @@ class ModelManager
 
             $mainModel->{$relation}()->save($mainModel->{$relation});
         } else {
-            $mainModel->{$relation}->delete();
+            if ($mainModel->{$relation}->exists) {
+                $mainModel->{$relation}->delete();
+            }
         }
     }
